[refactor] 03/25 커뮤니티 리스트 전용 스켈레톤 수정 #416
